storage: FichierPathStrategy

+++ igloo/igloo-components/storage-impl/src/main/java/org/igloo/storage/impl/StorageOperations.java
@@ -1,5 +1,6 @@
 package org.igloo.storage.impl;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -7,9 +8,12 @@ import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Path;
 
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.util.StringUtils;
 
 public class StorageOperations {
 
@@ -37,6 +41,15 @@ public class StorageOperations {
 	}
 
 	public void copy(InputStream inputStream, Path absolutePath) throws IOException {
+		Path dirPath = absolutePath.getParent();
+		File directory = dirPath.toFile();
+		if (directory.exists() && !directory.isDirectory()) {
+			throw new IllegalStateException(String.format("Parent path %s already exists and is not a directory", dirPath.toAbsolutePath()));
+		}
+		if (!directory.exists()) {
+			FileUtils.forceMkdir(directory);
+		}
+
 		try (FileOutputStream fos = new FileOutputStream(absolutePath.toString())) {
 			IOUtils.copy(inputStream, fos);
 		} catch (RuntimeException e) {
